(refactor): use getStripeInstance instead of importing stripe from npm in many places
