+++ imports/plugins/included/payments-stripe/server/no-meteor/util/stripeCreateRefund.js
@@ -1,7 +1,6 @@
-import stripeNpm from "stripe";
-import packageJson from "/package.json";
 import Logger from "@reactioncommerce/logger";
 import getStripeApiKey from "./getStripeApiKey";
+import getStripeInstance from "./getStripeInstance";
 import formatForStripe from "./formatForStripe";
 
 /**
@@ -18,12 +17,7 @@ export default async function stripeCreateRefund(context, paymentMethod, amount)
   let result;
   try {
     const stripeKey = await getStripeApiKey(context, paymentMethod.paymentPluginName, paymentMethod.shopId);
-    const stripe = stripeNpm(stripeKey);
-    stripe.setAppInfo({
-      name: "ReactionCommerce",
-      version: packageJson.version,
-      url: packageJson.url
-    });
+    const stripe = getStripeInstance(stripeKey);
 
     const refundResult = await stripe.refunds.create({ charge: paymentMethod.transactionId, amount: formatForStripe(amount) });
     Logger.debug(refundResult);
